Revert "refactor: simplify tunnel lifecycle and contracts"
This reverts commit aa2cd26a097f8c16a9a6aafcfbc4519c4c8686a6.

+++ src/main/java/com/huawei/devbridge/relaycontroller/domain/model/Tunnel.java
@@ -15,7 +15,7 @@ public class Tunnel {
     private String tunnelId;
     private Long tunnelCode;
     private String clusterId;
-    private Long expiration;
+    private Integer expiration;
     private String namespace;
     private String description;
     private Long bandwidthUsed;
